Update: 현재 날짜 값 가져오기
- 내역 추가 시 날짜 업데이트 안되는 문제 -> hooks로 현재 날짜 가져오도록 수정
